// web/src/api/generated/model/nodeType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.11.7
 */

export type NodeType = typeof NodeType[keyof typeof NodeType];


export const NodeType = {
  source: 'source',
  enricher: 'enricher',
  destination: 'destination',
  filter: 'filter',
  sorter: 'sorter',
  selector: 'selector',
  combiner: 'combiner',
} as const;